backend/requirements.txt
asgiref==3.10.0
brotli==1.2.0
click==8.3.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.0
sqlparse==0.5.3
uvicorn==0.38.0
whitenoise==6.11.0
